Finished all requirements for week 1
Added method on CaesarCipher that encrypts message using two keys.
Added class WordPlay that test if a character is a vowel.
Added method in WordPlay to replace a vowel with a given char
Added method in WordPlay that emphasizes a character depending on
if it index is odd or even in the String.
Finished implementing in Main final test cases of the end of week 1
quizz.