Changed some code implementations
-Modified the encrypt with two keys method on CaesarCipher to reduce the
lines of code required to encrypt a message using the two keys.

This method now encrypts the message using both keys and then replaces
them on a StringBuilder depending on if the index for the character is
even or odd.

-Fixed emphasize method on WordPlay in the same way, instead of
builder.append() it now uses builder.setCharAt and replaces the
character based on if the index is even or Odd.